refactor: enable better treeshaking
