+++ sources/advanced/entries/definitions.ts
@@ -1,0 +1,11 @@
+import {Command} from '../Command';
+
+/**
+         * A command that prints the clipanion definitions.
+         */
+export class DefinitionsCommand extends Command<any> {
+    static path = [[`--clipanion=definitions`]];
+    async execute() {
+        this.context.stdout.write(`${JSON.stringify(this.cli.definitions(), null, 2)}\n`);
+    }
+}
