+++ sources/advanced/entries/index.ts
@@ -1,0 +1,3 @@
+export * from './definitions'
+export * from './help'
+export * from './version'
