+++ sources/advanced/entries/version.ts
@@ -1,0 +1,13 @@
+import {Command} from '../Command';
+
+   /**
+         * A command that prints the version of the binary (`cli.binaryVersion`).
+         *
+         * Paths: `-v`, `--version`
+         */
+export class VersionCommand extends Command<any> {
+  static paths = [[`-v`], [`--version`]];
+  async execute() {
+    this.context.stdout.write(`${this.cli.binaryVersion ?? `<unknown>`}\n`);
+  }
+}
