+++ sources/advanced/index.ts
@@ -4,3 +4,7 @@ export {BaseContext, Cli, CliOptions} from './Cli';
 export {CommandClass, Usage, Definition} from './Command';
 
 export {UsageError} from '../errors';
+    
+export * as Entries from './entries'
+export * as Arguments from './arguments'
+
