+++ tests/advanced.test.ts
@@ -3,7 +3,7 @@ import chai, {expect}               from 'chai';
 import getStream                    from 'get-stream';
 import {PassThrough}                from 'stream';
 
-import {Cli, CommandClass, Command, CliOptions} from '../sources/advanced';
+import { Cli, CommandClass, Command, CliOptions, Arguments, Entries } from '../sources/advanced';
 
 chai.use(chaiAsPromised);
 
@@ -76,7 +76,7 @@ describe(`Advanced`, () => {
 
                 expect(cli.process([`-h`]).path).to.deep.equal([]);
 
-                cli.register(Command.Entries.Help);
+                cli.register(Entries.HelpCommand);
                 expect(cli.process([`-h`]).path).to.deep.equal([`-h`]);
 
                 expect(cli.process([`b`, `--help`]).path).to.deep.equal([`b`]);
@@ -85,7 +85,7 @@ describe(`Advanced`, () => {
 
             it(`should display the usage`, async () => {
                 const cli = new Cli()
-                cli.register(Command.Entries.Help);
+                cli.register(Entries.HelpCommand);
 
                 expect(await runCli(cli, [`-h`])).to.equal(cli.usage(null));
                 expect(await runCli(cli, [`--help`])).to.equal(cli.usage(null));
@@ -95,7 +95,7 @@ describe(`Advanced`, () => {
         describe(`version`, () => {
             it(`should display the version of the binary`, async () => {
                 const cli = new Cli({binaryVersion: `2.3.4`})
-                cli.register(Command.Entries.Version);
+                cli.register(Entries.VersionCommand);
 
                 expect(await runCli(cli, [`-v`])).to.equal(`2.3.4\n`);
                 expect(await runCli(cli, [`--version`])).to.equal(`2.3.4\n`);
@@ -104,7 +104,7 @@ describe(`Advanced`, () => {
 
             it(`should display "<unknown>" when no version is specified`, async () => {
                 const cli = new Cli()
-                cli.register(Command.Entries.Version);
+                cli.register(Entries.VersionCommand);
 
                 expect(await runCli(cli, [`-v`])).to.equal(`<unknown>\n`);
                 expect(await runCli(cli, [`--version`])).to.equal(`<unknown>\n`);
@@ -140,7 +140,7 @@ describe(`Advanced`, () => {
     it(`should print the help message when using --help`, async () => {
         const output = await runCli(() => {
             class CommandA extends Command {
-                foo = Command.Boolean(`--foo`);
+                foo = Arguments.Boolean(`--foo`);
                 async execute() {log(this)}
             }
 
@@ -157,7 +157,7 @@ describe(`Advanced`, () => {
     it(`shouldn't detect --help past the -- separator`, async () => {
         const output = await runCli(() => {
             class CommandA extends Command {
-                arg = Command.String();
+                arg = Arguments.String();
                 async execute() {log(this)}
             }
             return [
@@ -174,7 +174,7 @@ describe(`Advanced`, () => {
     it(`shouldn't detect --help on proxies`, async () => {
         const output = await runCli(() => {
             class CommandA extends Command {
-                args = Command.Proxy();
+                args = Arguments.Proxy();
                 async execute() {log(this, [`args`])}
             }
             return [
@@ -257,7 +257,7 @@ describe(`Advanced`, () => {
     it(`should support inheritance`, async () => {
         const output = await runCli(() => {
             abstract class CommandA extends Command {
-                foo = Command.String(`--foo`);
+                foo = Arguments.String(`--foo`);
                 abstract execute(): Promise<number | void>;
             }
 
@@ -277,9 +277,9 @@ describe(`Advanced`, () => {
 
     it(`derives positional argument names from the property name`, async () => {
         class CommandA extends Command {
-            workspaceName = Command.String();
-            extra = Command.String({required: false});
-            scriptName = Command.String();
+            workspaceName = Arguments.String();
+            extra = Arguments.String({required: false});
+            scriptName = Arguments.String();
 
             static paths = [[`workspace`]];
             async execute() {
@@ -294,7 +294,7 @@ describe(`Advanced`, () => {
 
     it(`derives rest argument names from the property name`, async () => {
         class CommandA extends Command {
-            workspaceNames = Command.Rest({required: 2});
+            workspaceNames = Arguments.Rest({required: 2});
 
             static paths = [[`clean`]];
             async execute() {
@@ -309,7 +309,7 @@ describe(`Advanced`, () => {
 
     it(`supports strings that act like booleans if not bound to a value`, async () => {
         class CommandA extends Command {
-            enableDebugger = Command.String(`--break`, false, {tolerateBoolean: true});
+            enableDebugger = Arguments.String(`--break`, false, {tolerateBoolean: true});
 
             async execute() {
                 log(this, [`enableDebugger`]);
@@ -317,7 +317,7 @@ describe(`Advanced`, () => {
         }
 
         class InvertedCommandA extends Command {
-            enableDebugger = Command.String(`--break`, true, {tolerateBoolean: true});
+            enableDebugger = Arguments.String(`--break`, true, {tolerateBoolean: true});
 
             async execute() {
                 log(this, [`enableDebugger`]);
@@ -354,7 +354,7 @@ describe(`Advanced`, () => {
 
     it(`shouldn't crash when throwing non-error exceptions`, async () => {
         class CommandA extends Command {
-            thisNameIsntUsed = Command.String({name: 'prettyName'});
+            thisNameIsntUsed = Arguments.String({name: 'prettyName'});
             async execute() {throw 42}
         }
 
@@ -456,7 +456,7 @@ describe(`Advanced`, () => {
 
     it(`should extract counter options from complex options`, async () => {
         class CommandA extends Command {
-            verbose = Command.Counter(`-v,--verbose`, 0);
+            verbose = Arguments.Counter(`-v,--verbose`, 0);
             async execute() {}
         }
 
@@ -479,9 +479,9 @@ describe(`Advanced`, () => {
 
     it(`should print flags with a description separately`, async () => {
         class CommandA extends Command {
-            verbose = Command.Boolean('--verbose', {description: 'Log output'});
-            output = Command.String('--output', {description: 'The output directory'});
-            message = Command.String('--message');
+            verbose = Arguments.Boolean('--verbose', {description: 'Log output'});
+            output = Arguments.String('--output', {description: 'The output directory'});
+            message = Arguments.String('--message');
 
             static paths = [[`greet`]];
             async execute() {
@@ -496,7 +496,7 @@ describe(`Advanced`, () => {
 
     it(`should support tuples`, async () => {
         class PointCommand extends Command {
-            point = Command.String(`--point`, {arity: 3});
+            point = Arguments.String(`--point`, {arity: 3});
             async execute() {}
         }
 
@@ -507,8 +507,8 @@ describe(`Advanced`, () => {
 
     it(`should extract tuples from complex options surrounded by rest arguments`, async () => {
         class PointCommand extends Command {
-            point = Command.String(`--point`, {arity: 3});
-            rest = Command.Rest();
+            point = Arguments.String(`--point`, {arity: 3});
+            rest = Arguments.Rest();
             async execute() {}
         }
 
@@ -524,7 +524,7 @@ describe(`Advanced`, () => {
 
     it(`should throw acceptable errors when tuple length is not finite`, async () => {
         class PointCommand extends Command {
-            point = Command.String(`--point`, {arity: Infinity});
+            point = Arguments.String(`--point`, {arity: Infinity});
             async execute() {}
         }
 
@@ -535,7 +535,7 @@ describe(`Advanced`, () => {
 
     it(`should throw acceptable errors when tuple length is not an integer`, async () => {
         class PointCommand extends Command {
-            point = Command.String(`--point`, {arity: 1.5});
+            point = Arguments.String(`--point`, {arity: 1.5});
             async execute() {}
         }
 
@@ -546,7 +546,7 @@ describe(`Advanced`, () => {
 
     it(`should throw acceptable errors when tuple length is not positive`, async () => {
         class PointCommand extends Command {
-            point = Command.String(`--point`, {arity: -1});
+            point = Arguments.String(`--point`, {arity: -1});
             async execute() {}
         }
 
@@ -557,7 +557,7 @@ describe(`Advanced`, () => {
 
     it(`should throw acceptable errors when not enough arguments are passed to a tuple`, async () => {
         class PointCommand extends Command {
-            point = Command.String(`--point`, {arity: 3});
+            point = Arguments.String(`--point`, {arity: 3});
             async execute() {}
         }
 
@@ -579,7 +579,7 @@ describe(`Advanced`, () => {
 
     it(`should extract string arrays from complex options`, async () => {
         class IncludeCommand extends Command {
-            include = Command.Array(`--include`, []);
+            include = Arguments.Array(`--include`, []);
             async execute() {}
         }
 
@@ -592,7 +592,7 @@ describe(`Advanced`, () => {
 
     it(`should extract tuple arrays from complex options`, async () => {
         class IncludeCommand extends Command {
-            position = Command.Array(`--position`, [], {arity: 3});
+            position = Arguments.Array(`--position`, [], {arity: 3});
             async execute() {}
         }
 
@@ -618,7 +618,7 @@ describe(`Advanced`, () => {
 
     it(`should support optional string positionals`, async () => {
         class ThingCommand extends Command {
-            thing = Command.String({required: false});
+            thing = Arguments.String({required: false});
             async execute() {}
         }
 
@@ -630,8 +630,8 @@ describe(`Advanced`, () => {
 
     it(`should support optional string positionals after required string positionals`, async () => {
         class CopyCommand extends Command {
-            requiredThing = Command.String();
-            optionalThing = Command.String({required: false});
+            requiredThing = Arguments.String();
+            optionalThing = Arguments.String({required: false});
             async execute() {}
         }
 
@@ -643,8 +643,8 @@ describe(`Advanced`, () => {
 
     it(`should support optional string positionals before required string positionals`, async () => {
         class CopyCommand extends Command {
-            optionalThing = Command.String({required: false});
-            requiredThing = Command.String();
+            optionalThing = Arguments.String({required: false});
+            requiredThing = Arguments.String();
             async execute() {}
         }
 
@@ -656,8 +656,8 @@ describe(`Advanced`, () => {
 
     it(`should support required positionals after rest arguments`, async () => {
         class CopyCommand extends Command {
-            sources = Command.Rest();
-            destination = Command.String();
+            sources = Arguments.Rest();
+            destination = Arguments.String();
             async execute() {}
         }
 
@@ -681,7 +681,7 @@ describe(`Advanced`, () => {
 
     it(`should support rest arguments with a minimum required length`, async () => {
         class CopyCommand extends Command {
-            sources = Command.Rest({required: 1});
+            sources = Arguments.Rest({required: 1});
             async execute() {}
         }
 
@@ -695,8 +695,8 @@ describe(`Advanced`, () => {
 
     it(`should support required positionals after rest arguments with a minimum required length`, async () => {
         class CopyCommand extends Command {
-            sources = Command.Rest({required: 1});
-            destination = Command.String();
+            sources = Arguments.Rest({required: 1});
+            destination = Arguments.String();
             async execute() {}
         }
 
@@ -720,10 +720,10 @@ describe(`Advanced`, () => {
     // We have this in the README, that's why we're testing it
     it(`should support implementing a cp-like command`, async () => {
         class CopyCommand extends Command {
-            sources = Command.Rest({required: 1});
-            destination = Command.String();
-            force = Command.Boolean(`-f,--force`, false);
-            reflink = Command.String(`--reflink`, false, {tolerateBoolean: true});
+            sources = Arguments.Rest({required: 1});
+            destination = Arguments.String();
+            force = Arguments.Boolean(`-f,--force`, false);
+            reflink = Arguments.String(`--reflink`, false, {tolerateBoolean: true});
             async execute() {}
         }
 
@@ -776,7 +776,7 @@ describe(`Advanced`, () => {
 
     it(`should support proxies`, async () => {
         class CopyCommand extends Command {
-            args = Command.Proxy();
+            args = Arguments.Proxy();
             async execute() {}
         }
 
@@ -790,7 +790,7 @@ describe(`Advanced`, () => {
 
     it(`should support proxies with a minimum required length`, async () => {
         class CopyCommand extends Command {
-            args = Command.Proxy({required: 1})
+            args = Arguments.Proxy({required: 1})
             async execute() {}
         }
 
@@ -804,7 +804,7 @@ describe(`Advanced`, () => {
 
     it(`should not allow negating options with arity 1 that already start with "--no-"`, async () => {
         class FooCommand extends Command {
-            noRedacted = Command.Boolean(`--no-redacted`);
+            noRedacted = Arguments.Boolean(`--no-redacted`);
             async execute() {}
         }
